Adds MiqServerBaseMethods
Moves common methods out of MiqServer to be reused in a later commit.
