+++ app/models/miq_server.rb
@@ -1,5 +1,6 @@
 require 'resolv'
 require 'miq_server/base_constants'
+require 'miq_server/base_methods'
 
 class MiqServer < ApplicationRecord
   include_concern 'WorkerManagement'
@@ -25,8 +26,6 @@ class MiqServer < ApplicationRecord
   has_many                :messages,  :as => :handler, :class_name => 'MiqQueue'
   has_many                :miq_events, :as => :target, :dependent => :destroy
 
-  cattr_accessor          :my_guid_cache
-
   before_destroy          :validate_is_deleteable
 
   default_value_for       :rhn_mirror, false
@@ -34,6 +33,7 @@ class MiqServer < ApplicationRecord
   virtual_column :zone_description, :type => :string
 
   include MiqServerBaseConstants
+  include MiqServerBaseMethods
 
   def self.active_miq_servers
     where(:status => STATUSES_ACTIVE)
@@ -89,15 +89,6 @@ class MiqServer < ApplicationRecord
     # create root data directory
     data_dir = File.join(File.expand_path(Rails.root), "data")
     Dir.mkdir data_dir unless File.exist?(data_dir)
-  end
-
-  def self.pidfile
-    @pidfile ||= "#{Rails.root}/tmp/pids/evm.pid"
-  end
-
-  def self.running?
-    p = PidFile.new(pidfile)
-    p.running? ? p.pid : false
   end
 
   def start
@@ -361,57 +352,6 @@ class MiqServer < ApplicationRecord
     end
   end
 
-  def stop(sync = false)
-    return if self.stopped?
-
-    shutdown_and_exit_queue
-    wait_for_stopped if sync
-  end
-
-  def wait_for_stopped
-    loop do
-      reload
-      break if self.stopped?
-      sleep stop_poll
-    end
-  end
-
-  def self.stop(sync = false)
-    svr = my_server(true) rescue nil
-    svr.stop(sync) unless svr.nil?
-    PidFile.new(pidfile).remove
-  end
-
-  def kill
-    # Kill all the workers of this server
-    kill_all_workers
-
-    # Then kill this server
-    _log.info("initiated for #{format_full_log_msg}")
-    update_attributes(:stopped_on => Time.now.utc, :status => "killed", :is_master => false)
-    (pid == Process.pid) ? shutdown_and_exit : Process.kill(9, pid)
-  end
-
-  def kill_all_workers
-    return unless is_local?
-
-    killed_workers = []
-    miq_workers.each do |w|
-      next unless MiqWorker::STATUSES_CURRENT_OR_STARTING.include?(w.status)
-
-      w.kill
-      worker_delete(w.pid)
-      killed_workers << w
-    end
-    miq_workers.delete(*killed_workers) unless killed_workers.empty?
-  end
-
-  def self.kill
-    svr = my_server(true)
-    svr.kill unless svr.nil?
-    PidFile.new(pidfile).remove
-  end
-
   def shutdown
     _log.info("initiated for #{format_full_log_msg}")
     MiqEvent.raise_evm_event(self, "evm_server_stop")
@@ -465,14 +405,6 @@ class MiqServer < ApplicationRecord
     ArApplicationName.name = database_application_name
   end
 
-  def is_local?
-    guid == MiqServer.my_guid
-  end
-
-  def is_remote?
-    !is_local?
-  end
-
   def is_recently_active?
     last_heartbeat && (last_heartbeat >= 10.minutes.ago.utc)
   end
@@ -500,10 +432,6 @@ class MiqServer < ApplicationRecord
     status == "started"
   end
 
-  def stopped?
-    STATUSES_STOPPED.include?(status)
-  end
-
   def active?
     STATUSES_ACTIVE.include?(status)
   end
@@ -526,19 +454,6 @@ class MiqServer < ApplicationRecord
     message = "Waiting for #{wcnt} #{workers} to start"
     result.merge(:message => message)
   end
-
-  #
-  # Zone and Role methods
-  #
-  def self.my_guid
-    @@my_guid_cache ||= begin
-      guid_file = Rails.root.join("GUID")
-      File.write(guid_file, SecureRandom.uuid) unless File.exist?(guid_file)
-      File.read(guid_file).strip
-    end
-  end
-
-  cache_with_timeout(:my_server) { find_by(:guid => my_guid) }
 
   def self.my_zone(force_reload = false)
     my_server(force_reload).my_zone
